delete component complete

+++ server/routes/shelf.router.js
@@ -34,9 +34,9 @@ router.post('/',rejectUnauthenticated, (req, res) => {
 router.delete('/:id', rejectUnauthenticated, (req, res) => {
   console.log(req.params.id);
   let id = req.params.id;
-  let queryText = `DELETE FROM "item" WHERE "id" = $1`;
+  let queryText = `DELETE FROM "item" WHERE "id" = $1 AND "user_id" = $2`;
 
-  pool.query(queryText, [id])
+  pool.query(queryText, [id, req.user.id])
     .then(result => {
       res.sendStatus(204);
     })
